added a test for fatal error discord logging
